+++ lib/src/extensions/iterable_extensions.dart
@@ -16,6 +16,7 @@ extension _IterableIndexing<E> on Iterable<E> {
   }
 }
 
+
 /// Conversion helpers for non-null [Iterable] collections.
 extension IterableConversionX<E> on Iterable<E> {
   E? _valueAt(int index) => _elementAtOrNull(index);
@@ -177,11 +178,13 @@ extension IterableConversionX<E> on Iterable<E> {
     dynamic innerMapKey,
     int? innerIndex,
     List<T>? defaultValue,
+    ElementConverter<T>? elementConverter,
   }) => ConvertObjectImpl.toList<T>(
     _valueAt(index),
     mapKey: innerMapKey,
     listIndex: innerIndex,
     defaultValue: defaultValue,
+    elementConverter: elementConverter,
     debugInfo: {'index': index},
   );
 
@@ -191,11 +194,13 @@ extension IterableConversionX<E> on Iterable<E> {
     dynamic innerMapKey,
     int? innerIndex,
     Set<T>? defaultValue,
+    ElementConverter<T>? elementConverter,
   }) => ConvertObjectImpl.toSet<T>(
     _valueAt(index),
     mapKey: innerMapKey,
     listIndex: innerIndex,
     defaultValue: defaultValue,
+    elementConverter: elementConverter,
     debugInfo: {'index': index},
   );
 
@@ -205,11 +210,15 @@ extension IterableConversionX<E> on Iterable<E> {
     dynamic innerMapKey,
     int? innerIndex,
     Map<K2, V2>? defaultValue,
+    ElementConverter<K2>? keyConverter,
+    ElementConverter<V2>? valueConverter,
   }) => ConvertObjectImpl.toMap<K2, V2>(
     _valueAt(index),
     mapKey: innerMapKey,
     listIndex: innerIndex,
     defaultValue: defaultValue,
+    keyConverter: keyConverter,
+    valueConverter: valueConverter,
     debugInfo: {'index': index},
   );
 
@@ -220,14 +229,22 @@ extension IterableConversionX<E> on Iterable<E> {
     dynamic innerMapKey,
     int? innerIndex,
     T? defaultValue,
-  }) => ConvertObjectImpl.toEnum<T>(
-    _valueAt(index),
-    parser: parser,
-    mapKey: innerMapKey,
-    listIndex: innerIndex,
-    defaultValue: defaultValue,
-    debugInfo: {'index': index},
-  );
+    Map<String, dynamic>? debugInfo,
+  }) {
+    final info = <String, dynamic>{};
+    if (debugInfo != null && debugInfo.isNotEmpty) {
+      info.addAll(debugInfo);
+    }
+    info['index'] = index;
+    return ConvertObjectImpl.toEnum<T>(
+      _valueAt(index),
+      parser: parser,
+      mapKey: innerMapKey,
+      listIndex: innerIndex,
+      defaultValue: defaultValue,
+      debugInfo: info,
+    );
+  }
 
   // Convert all
   /// Converts every element in this iterable to [T].
@@ -495,11 +512,13 @@ extension NullableIterableConversionX<E> on Iterable<E>? {
     dynamic innerMapKey,
     int? innerIndex,
     List<T>? defaultValue,
+    ElementConverter<T>? elementConverter,
   }) => ConvertObjectImpl.tryToList<T>(
     _firstForIndices(index, alternativeIndices: alternativeIndices),
     mapKey: innerMapKey,
     listIndex: innerIndex,
     defaultValue: defaultValue,
+    elementConverter: elementConverter,
     debugInfo: {
       'index': index,
       if (alternativeIndices != null && alternativeIndices.isNotEmpty)
@@ -514,11 +533,13 @@ extension NullableIterableConversionX<E> on Iterable<E>? {
     dynamic innerMapKey,
     int? innerIndex,
     Set<T>? defaultValue,
+    ElementConverter<T>? elementConverter,
   }) => ConvertObjectImpl.tryToSet<T>(
     _firstForIndices(index, alternativeIndices: alternativeIndices),
     mapKey: innerMapKey,
     listIndex: innerIndex,
     defaultValue: defaultValue,
+    elementConverter: elementConverter,
     debugInfo: {
       'index': index,
       if (alternativeIndices != null && alternativeIndices.isNotEmpty)
@@ -533,11 +554,15 @@ extension NullableIterableConversionX<E> on Iterable<E>? {
     dynamic innerMapKey,
     int? innerIndex,
     Map<K2, V2>? defaultValue,
+    ElementConverter<K2>? keyConverter,
+    ElementConverter<V2>? valueConverter,
   }) => ConvertObjectImpl.tryToMap<K2, V2>(
     _firstForIndices(index, alternativeIndices: alternativeIndices),
     mapKey: innerMapKey,
     listIndex: innerIndex,
     defaultValue: defaultValue,
+    keyConverter: keyConverter,
+    valueConverter: valueConverter,
     debugInfo: {
       'index': index,
       if (alternativeIndices != null && alternativeIndices.isNotEmpty)
@@ -553,18 +578,25 @@ extension NullableIterableConversionX<E> on Iterable<E>? {
     dynamic innerMapKey,
     int? innerIndex,
     T? defaultValue,
-  }) => ConvertObjectImpl.tryToEnum<T>(
-    _firstForIndices(index, alternativeIndices: alternativeIndices),
-    parser: parser,
-    mapKey: innerMapKey,
-    listIndex: innerIndex,
-    defaultValue: defaultValue,
-    debugInfo: {
-      'index': index,
-      if (alternativeIndices != null && alternativeIndices.isNotEmpty)
-        'altIndexes': alternativeIndices,
-    },
-  );
+    Map<String, dynamic>? debugInfo,
+  }) {
+    final info = <String, dynamic>{};
+    if (debugInfo != null && debugInfo.isNotEmpty) {
+      info.addAll(debugInfo);
+    }
+    info['index'] = index;
+    if (alternativeIndices != null && alternativeIndices.isNotEmpty) {
+      info['altIndexes'] = alternativeIndices;
+    }
+    return ConvertObjectImpl.tryToEnum<T>(
+      _firstForIndices(index, alternativeIndices: alternativeIndices),
+      parser: parser,
+      mapKey: innerMapKey,
+      listIndex: innerIndex,
+      defaultValue: defaultValue,
+      debugInfo: info,
+    );
+  }
 }
 
 /// Converts nullable sets into a [Set] of a different type.